Split up the tests
